+++ 0x1E-search_algorithms/1-binary.c
@@ -7,16 +7,18 @@
  *
  * return: void
  */
-void print_array(int *array, size_t size) {
-    size_t i;
+void print_array(int *array, size_t size)
+{
+	size_t i;
 
-    printf("Searching in array:");
-    for (i = 0; i < size; i++) {
-        printf(" %d", array[i]);
-        if (i != size - 1)
-            printf(",");
-    }
-    printf("\n");
+	printf("Searching in array:");
+	for (i = 0; i < size; i++)
+	{
+		printf(" %d", array[i]);
+		if (i != size - 1)
+			printf(",");
+	}
+	printf("\n");
 }
 
 /**
@@ -28,25 +30,28 @@ void print_array(int *array, size_t size) {
  *
  * Return: index where value is located, or -1 on failure
  */
-int binary_search(int *array, size_t size, int value) {
-    size_t low, mid, high;
+int binary_search(int *array, size_t size, int value)
+{
+	size_t low, mid, high;
 
-    if (array != NULL && size > 0) {
-        low = 0;
-        high = size - 1;
+	if (array != NULL && size > 0)
+	{
+		low = 0;
+		high = size - 1;
 
-        print_array(array + low, high + 1 - low);
+		print_array(array + low, high + 1 - low);
 
-        while (low < high) {
-            mid = (low + high) / 2;
-            if (array[mid] < value)
-                low = mid + 1;
-            else if (array[mid] > value)
-                high = mid;
-            else
-                return (int) (mid);
-            print_array(array + low, high + 1 - low);
-        }
-    }
-    return (-1);
+		while (low < high)
+		{
+			mid = (low + high) / 2;
+			if (array[mid] < value)
+				low = mid + 1;
+			else if (array[mid] > value)
+				high = mid;
+			else
+				return (mid);
+			print_array(array + low, high + 1 - low);
+		}
+	}
+	return (-1);
 }
