+++ 0x1E-search_algorithms/102-interpolation.c
@@ -9,26 +9,30 @@
  *
  * Return: the first index where value is located, or -1 on failure
  */
-int interpolation_search(int *array, size_t size, int value) {
-    size_t l, m, r;
+int interpolation_search(int *array, size_t size, int value)
+{
+	size_t l, m, r;
 
-    if (array != NULL && size > 0) {
-        l = 0;
-        r = size - 1;
-        while (array[l] != array[r]) {
-            m = l + (((double) (r - l) / (array[r] - array[l])) * (value - array[l]));
-            if (m < l || m > r) {
-                printf("Value checked array[%zu] is out of range\n", m);
-                break;
-            }
-            printf("Value checked array[%zu] = [%d]\n", m, array[m]);
-            if (array[m] < value)
-                l = m + 1;
-            else if (array[m] > value)
-                r = m - 1;
-            else
-                return (int) (m);
-        }
-    }
-    return (-1);
+	if (array != NULL && size > 0)
+	{
+		l = 0;
+		r = size - 1;
+		while (array[l] != array[r])
+		{
+			m = l + (((double) (r - l) / (array[r] - array[l])) * (value - array[l]));
+			if (m < l || m > r)
+			{
+				printf("Value checked array[%zu] is out of range\n", m);
+				break;
+			}
+			printf("Value checked array[%zu] = [%d]\n", m, array[m]);
+			if (array[m] < value)
+				l = m + 1;
+			else if (array[m] > value)
+				r = m - 1;
+			else
+				return (m);
+		}
+	}
+	return (-1);
 }
